Very basic drag drop
TLDR: Pushing first so that I can work on remote machine.

Was trying to get EventSystem and Canvas to work but gave up, settled with basic click and drag system.

Will finish up the cup system. Temporarily for milestone 1 will need to display text to show contents because haven't completed drawing for brewing part.

To say that this isn't milestone 1 worthy is semi true but, then again, it's a click and drag game. Once we can get the click and drag components work, it's just the matter of duplicating to different ingredients and equipment.

Need to discuss and figure out camera